end moving EncryptionKey to the protocol
