+++ client/Cargo.toml
@@ -9,7 +9,6 @@ license = "MIT OR Apache-2.0"
 rammingen_protocol = { path = "../protocol" }
 rammingen_sdk = { path = "../sdk" }
 
-aes-siv.workspace = true
 anyhow.workspace = true
 base64.workspace = true
 byteorder.workspace = true
